src/validation/validators/email/email-validation.spec.ts
import { InvalidFiedlError } from '@/validation/errors'
import { EmailValidation } from './email-validation'
import faker from 'faker'

const makeSut = (): EmailValidation =>
  new EmailValidation(faker.database.column())
describe('EmailValidation', () => {
  test('Should return error if the email is invalid', () => {
    const sut = makeSut()
    const error = sut.validate(faker.random.word())
    expect(error).toEqual(new InvalidFiedlError())
  })

  test('Should return false if the email is valid', () => {
    const sut = makeSut()
    const error = sut.validate(faker.internet.email())
    expect(error).toBeFalsy()
  })

  test('Should return false if the email is empty', () => {
    const sut = makeSut()
    const error = sut.validate('')
    expect(error).toBeFalsy()
  })
})
